Program 3: TwentyFour Part Two
Program 3: TwentyFour Part Two
Description
24 is a math game, in which the player is presented with 4 numbers and they must determine a way
to combine those numbers using mathematical operations to get 24 as the result. Each number must
be used, and each number can be used only once.
As an example, if the user is presented with the card shown below:
they may come up with the following solution (which may not be the only one!):
5 x 5 = 25
4 – 3 = 1
25 – 1 = 24
See https://www.24game.com/t-about-howtoplay.aspx for more information.
In Program 1, you developed a C program that allows the user to play a simplified version of this
game, where calculations were done in order. For this program, you will allow the user to enter a full
expression for their solution, which your program will then evaluate. This means that if the user is
presented with the example above, they may type something like (5 * 5) - (4 - 3) for their
solution, and your program will evaluate it as shown above!
The four numbers presented in each puzzle are all single digits (numbers 1-9). The user does not
need to use all four mathematical operations.